AUDIT REFACTOR 20: Simplify permit API and update function signatures

+++ test/TestUtils.t.sol
@@ -95,7 +95,8 @@ contract TestUtilsTest is Test {
             amountDelta: 500
         });
 
-        IPermit3.ChainPermits memory chainPermits = IPermit3.ChainPermits({ chainId: block.chainid, permits: permits });
+        IPermit3.ChainPermits memory chainPermits =
+            IPermit3.ChainPermits({ chainId: uint64(block.chainid), permits: permits });
 
         // Hash using utility
         bytes32 hash = Permit3TestUtils.hashChainPermits(permit3, chainPermits);
@@ -120,7 +121,7 @@ contract TestUtilsTest is Test {
 
     function test_hashEmptyChainPermits() public view {
         // Test hashEmptyChainPermits function
-        uint256 chainId = block.chainid;
+        uint64 chainId = uint64(block.chainid);
         bytes32 emptyHash = Permit3TestUtils.hashEmptyChainPermits(permit3, chainId);
 
         // Create empty permits manually
@@ -142,7 +143,7 @@ contract TestUtilsTest is Test {
             Permit3TestUtils.createTransferPermit(testToken, testRecipient, testAmount);
 
         // Verify the permit structure
-        assertEq(transferPermit.chainId, block.chainid);
+        assertEq(transferPermit.chainId, uint64(block.chainid));
         assertEq(transferPermit.permits.length, 1);
         assertEq(transferPermit.permits[0].modeOrExpiration, 0); // Transfer mode
         assertEq(transferPermit.permits[0].token, testToken);
